mqtt-server-novinar/requirements.txt
paho-mqtt
torch
torchvision
Pillow

